Add GHC-7.10 support
